XXX Total failure.
This is not going to work, as originally envisioned. I had imagined that
the MonoValue could be used as a post-search filter, but I see now that
this cannot work: the test predicate has to be applied during search to
reject undesirable matches. So the current design here is a failure,
which explains why Claude seems unable to port the unit test correctly:
I've asked it to perform a possible task.