Track variables in Always clauses.
This is the first in a long series of commits to fix AlwaysLink,
as described in the README-Algorithm.md file. I' mnot sure this edit
is correct; I thin the variables will have to travel with Parts,
or bridge across them. So this is a provisional first draft. Well
see how it goes.